--- js/chinhsuathongtin.js
@@ -1,1 +1,0 @@
-
