+++ src/chinhsuakykinh.html
@@ -14,7 +14,7 @@
 	<script language="JavaScript" type="text/javascript" src="../js/jquery.min.js" charset="utf-8"></script>
 	<script language="JavaScript" type="text/javascript" src="../js/bootstrap.js" charset="utf-8"></script>
 	<script src="../node_modules/pouchdb/dist/pouchdb.js"></script>
-	<script language="JavaScript" type="text/javascript" src="../js/chinhsuakykinh.js" charset="utf-8"></script>
+	<script language="JavaScript" type="text/javascript" src="../db/chinhsuakykinh.js" charset="utf-8"></script>
 
 	<script>if (window.module) module = window.module;</script>
 	<link href="../css/jquery-ui-1.10.1.css" rel="stylesheet">
