+++ src/home.html
@@ -68,10 +68,8 @@
 
 				<div id='calendar'></div>
 			</div>
-<<<<<<< HEAD
 
-=======
-			
+
 	</div>
 	<script>
 
@@ -86,10 +84,9 @@
 				location.reload();
 			}
 		});
->>>>>>> database-added
 
 		});
-		
+
 	</script>
 	<script type="text/javascript" src="../db/home.js"></script>
 </body>
